filters refactor
capability-based approach
separated files
