+++ lib/discord/src/design.ts
@@ -1,7 +1,5 @@
 import { lookupModule } from '@revenge-mod/modules/finders'
 import {
-    or,
-    preferExports,
     withDependencies,
     withProps,
     withSingleProp,
@@ -22,23 +20,21 @@ export let Design: Design = proxify(
         // ID: 3236
         // [3237, 1366, 3238, 3239, 2, ...];
         const [module] = lookupModule(
-            preferExports(
-                withProps<Design>('TableRow', 'Button'),
-                or(
-                    withDependencies(
-                        loose([
-                            null,
-                            null,
-                            null,
-                            relative.withDependencies([2], 1),
-                            2,
-                            null, // 3009
-                            null, // 3010
-                            null, // 3011
-                            null, // 3012
-                            relative.withDependencies([2], 2),
-                        ]),
-                    ),
+            withProps<Design>('TableRow', 'Button').and(
+                withDependencies(
+                    loose([
+                        null,
+                        null,
+                        null,
+                        relative.withDependencies([2], 1),
+                        2,
+                        null, // 3009
+                        null, // 3010
+                        null, // 3011
+                        null, // 3012
+                        relative.withDependencies([2], 2),
+                    ]),
+                ).or(
                     // TODO(PalmDevs): Remove once stable channel is > 295203 (for 295203 and below)
                     withDependencies(
                         loose([relative(1), null, relative(2), relative(3), 2]),
@@ -65,10 +61,9 @@ export let FormSwitch: DiscordModules.Components.FormSwitch = proxify(() => {
     // Deps: [47, 48, 35, 49]
 
     const [module] = lookupModule(
-        preferExports(
-            withSingleProp<{
-                FormSwitch: DiscordModules.Components.FormSwitch
-            }>('FormSwitch'),
+        withSingleProp<{
+            FormSwitch: DiscordModules.Components.FormSwitch
+        }>('FormSwitch').and(
             withDependencies([
                 [relative(1), relative(2), null, relative(3)],
                 ReactModuleId,
