+++ lib/externals/src/shopify.ts
@@ -1,6 +1,5 @@
 import { lookupModule } from '@revenge-mod/modules/finders'
 import {
-    preferExports,
     withDependencies,
     withProps,
 } from '@revenge-mod/modules/finders/filters'
@@ -14,8 +13,9 @@ import { proxify } from '@revenge-mod/utils/proxy'
 export let FlashList: typeof import('@shopify/flash-list') = proxify(
     () => {
         const [module] = lookupModule(
-            preferExports(
-                withProps<typeof FlashList>('FlashList'),
+            withProps<typeof FlashList>('FlashList').and(
+                // Dependencies. One in brackets are dynamic or late initialized.
+                // [React, RN, JSXRuntime, (FlashListExports), (Reanimated), (RNBottomSheet), ImportTracker, (BottomSheetFlashList)]
                 withDependencies([
                     ReactModuleId,
                     ReactNativeModuleId,
@@ -26,8 +26,6 @@ export let FlashList: typeof import('@shopify/flash-list') = proxify(
                     2,
                     null,
                 ]),
-                // Dependencies. One in brackets are dynamic or late initialized.
-                // [React, RN, JSXRuntime, (FlashListExports), (Reanimated), (RNBottomSheet), ImportTracker, (BottomSheetFlashList)]
             ),
             {
                 uninitialized: true,
